requirements.txt
trendspy>=0.1.6
praw>=7.7.0
python-dotenv>=1.0.0
requests>=2.31.0
scrapling>=0.3.14
biopython>=1.83
feedparser>=6.0.10
Jinja2>=3.1.0
